Reuse group-column controller for websocket logic
