+++ app/views/visualizations/_column.html.erb
@@ -2,11 +2,11 @@
 <li class="max-h-full w-64 flex flex-col shrink-0 p-2 border bg-body-contrast border-background-100 rounded-md shadow-xs cpy-grouping js-sortable-container <%= animation_css %>"
   style="animation-duration: 240ms;"
   id="<%= dom_id(grouping) %>"
-  data-controller="grouping-column visualization--board--column"
+  data-controller="grouping-column"
   data-grouping-column-scroll-to-on-connect-value="<%= local_assigns[:recently_created] == true %>"
+  data-grouping-column-grouping-id-value="<%= grouping.id %>"
   data-action="keypress.n@window->grouping-column#nWasPressed"
   data-visualization--board-target="column"
-  data-visualization--board--column-grouping-id-value="<%= grouping.id %>"
   >
   <header class="flex flex-row flex-nowrap gap-x-1 mb-4 px-2 relative" data-controller="dropdown">
     <%= link_to edit_visualization_grouping_path(grouping.visualization_id, grouping.id), class: "grow font-semibold truncate text-readable-content-500 hover:text-readable-content-700", data: { turbo_frame: 'grouping_form' } do %>
@@ -57,7 +57,7 @@
     data-sortable-shared-group-value="cards"
     data-sortable-move-path-value="<%= move_allocations_path %>"
     data-sortable-grouping-id="<%= grouping.id %>"
-    data-visualization--board--column-target="cardContainer"
+    data-grouping-column-target="cardContainer"
     >
 
     <% grouping.issues.each do |issue| %>
